web_app/requirements.txt
Flask==2.3.3
google-generativeai==0.8.5
pandas==2.2.2
Pillow==10.4.0
python-dotenv==1.0.0
gunicorn==22.0.0